Moved validate user to WebhookService
